backend/requirements.txt
# Core Dependencies
torch>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
pandas>=1.3.0

# SUMO Interface
traci>=1.18.0
sumolib>=1.18.0

# Backend (Optional for dashboard)
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
Flask>=2.0.0
Flask-CORS>=3.0.10
python-dotenv>=0.21.0

# Utilities
pyyaml>=6.0
tensorboard>=2.12.0
tqdm>=4.65.0